pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "FixaTons"
version = "1.0.0"
description = "A Python toolkit for eye-tracking datasets, saliency maps, and scanpath similarity metrics"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
authors = [{ name = "Dario Zanca", email = "dariozanca@gmail.com" }]
maintainers = [{ name = "Dario Zanca", email = "dariozanca@gmail.com" }]
keywords = ["eye-tracking", "fixations", "scanpath", "saliency", "visual-attention"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "numpy>=1.21.0",
  "pandas>=1.3.0",
  "opencv-python>=4.5.0",
  "matplotlib>=3.4.0",
  "nltk>=3.7",
  "multimatch>=0.1.0",
  "scipy>=1.7.0",
  "Pillow>=8.0.0",
]

[project.optional-dependencies]
dev = [
  "black>=22.0.0",
  "flake8>=4.0.0",
  "isort>=5.10.0",
  "mypy>=0.950",
  "pre-commit>=2.17.0",
  "pytest>=6.0.0",
  "pytest-cov>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/dariozanca/FixaTons"
Repository = "https://github.com/dariozanca/FixaTons"
Issues = "https://github.com/dariozanca/FixaTons/issues"
Documentation = "https://github.com/dariozanca/FixaTons#readme"

[project.scripts]
fixatons-datasets = "FixaTons.cli:main"

[tool.setuptools.packages.find]
include = ["FixaTons", "FixaTons.*"]

[tool.setuptools.package-data]
FixaTons = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["FixaTons"]
include_trailing_comma = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = ["cv2.*", "multimatch.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
